% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Remove_Zeros.R
\name{rm_0s_by_col}
\alias{rm_0s_by_col}
\title{To remove columns of 0s}
\usage{
rm_0s_by_col(x)
}
\arguments{
\item{x}{The table of numbers, must be a numeric or an integer vector.}
}
\value{
A numeric or integer table, with the table of all columns of 0s removed
}
\description{
These functions run through the columns of a set of data
to check for columns that contain only 0s
}
