% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Remove_Zeros.R
\name{rm_0s_by_row}
\alias{rm_0s_by_row}
\title{To remove rows of 0s}
\usage{
rm_0s_by_row(data = v)
}
\arguments{
\item{v}{The table of numbers, must be a numeric or an integer vector.}
}
\value{
A numeric or integer table, with the table of all rows of 0s removed
}
\description{
These functions run through the rows of a set of data
to check for rows that contain only 0s
}
